refactor(ui): extract add task modal component
